% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castAllStringsToNA}
\alias{castAllStringsToNA}
\title{castAllStringsToNA}
\usage{
castAllStringsToNA(dataset, excludeColumns = c())
}
\arguments{
\item{dataset}{dataframe}

\item{excludeColumns}{character}
}
\value{
dataframe
}
\description{
Removes all strings/words from specified columns in dataset
}
