% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_file_header}
\alias{generate_file_header}
\title{Generate a File Header}
\usage{
generate_file_header(dataset)
}
\arguments{
\item{dataset}{The input data frame.}
}
\value{
A data frame containing original and remapped column names.
}
\description{
This function generates a fileHeader object from a given data frame
which includes original names and remapped names of the data frame columns.
}
