% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunautLAIV-data.R
\docType{data}
\name{immunautLAIV}
\alias{immunautLAIV}
\title{Demo data set from immunaut package.
This data is used in this package examples.}
\format{
An object of class \code{data.frame} with 244 rows and 32 columns.
}
\usage{
data(immunautLAIV)
}
\description{
Demo data set from immunaut package.
This data is used in this package examples.
}
\examples{
\dontrun{
	data(immunautLAIV)
	## define settings variable
	settings <- list()
	settings$fileHeader <- generate_file_header(immunautLAIV)
	# ... and other settings
	results <- immunaut(immunautLAIV, settings)
}

}
\keyword{datasets}
