% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_var_empty}
\alias{is_var_empty}
\title{Check if request variable is Empty}
\usage{
is_var_empty(variable)
}
\arguments{
\item{variable}{The variable to check.}
}
\value{
boolean TRUE if the variable is considered empty, FALSE otherwise.
}
\description{
Checks if the given variable is empty and optionally logs the variable name.
}
\keyword{internal}
