% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNumeric}
\alias{isNumeric}
\title{Is Numeric}
\usage{
isNumeric(x)
}
\arguments{
\item{x}{Variable to be checked}
}
\value{
Logical indicating whether x is numeric and non-NA
}
\description{
Determines whether a variable is a number or a numeric string
}
\keyword{internal}
