% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pick_best_cluster_modularity}
\alias{pick_best_cluster_modularity}
\title{Pick Best Cluster by Modularity}
\usage{
pick_best_cluster_modularity(tsne_clust)
}
\arguments{
\item{tsne_clust}{A list of clustering results where each element contains clustering information,
including the modularity score.}
}
\value{
Returns the clustering result with the highest modularity score.
}
\description{
This function selects the best cluster from a list of clustering results
based on the highest modularity score.
}
\details{
The function iterates over a list of clustering results (\code{tsne_clust}) and
selects the cluster with the highest modularity score. If no clusters are valid or
the \code{tsne_clust} list is empty, the function will stop and return an error.
}
\keyword{internal}
