% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pick_best_cluster_silhouette}
\alias{pick_best_cluster_silhouette}
\title{Pick Best Cluster by Silhouette Score}
\usage{
pick_best_cluster_silhouette(tsne_clust)
}
\arguments{
\item{tsne_clust}{A list of clustering results where each element contains clustering information,
including the average silhouette score.}
}
\value{
Returns the clustering result with the highest average silhouette score.
}
\description{
This function selects the best cluster from a list of clustering results
based on the highest average silhouette score.
}
\details{
The function iterates over a list of clustering results (\code{tsne_clust}) and
selects the cluster with the highest average silhouette score. If no clusters are valid or
the \code{tsne_clust} list is empty, the function will stop and return an error.
}
\keyword{internal}
