% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_simon_ml}
\alias{auto_simon_ml}
\title{Auto ML Model Building with Caret}
\usage{
auto_simon_ml(
  dataset_ml,
  outcome,
  selectedPackages,
  selectedPartitionSplit = 0.7
)
}
\arguments{
\item{dataset_ml}{A data frame containing the dataset for training. All columns except the outcome
column should contain the features.}

\item{outcome}{A string specifying the name of the outcome column in the dataset. The outcome must
be a factor with two or more levels.}

\item{selectedPackages}{A character vector specifying the caret machine learning models to train.
For example, \code{c("nb", "rpart")}.}

\item{selectedPartitionSplit}{A numeric value between 0 and 1 specifying the proportion of the dataset
to use for training. Defaults to 0.7 (70\%).}
}
\value{
A list where each element is a trained model corresponding to the algorithms specified in
\code{selectedPackages}. Each element contains:
\itemize{
\item{\code{info}}: General information about the model, including resampling indices, problem type,
and outcome mapping.
\item{\code{training}}: The trained model and variable importance.
\item{\code{predictions}}: Predictions on the test set, including processed probabilities, confusion
matrix, and performance metrics such as AUROC and prAUC (if applicable).
}
}
\description{
This function automates the process of building machine learning models using the caret package.
It supports both binary and multi-class classification and trains models based on a user-specified
list of machine learning algorithms. The function splits the dataset into training and testing sets,
trains the models using cross-validation, and computes performance metrics such as confusion matrix,
AUROC (for binary classification), and prAUC (for binary classification).
}
\details{
The function handles binary and multi-class classification problems and trains models using the caret
package. It performs preprocessing (scaling, centering, and imputation of missing values), trains models
using cross-validation, and calculates performance metrics including confusion matrix, post-resampling,
and AUROC for binary classification problems.

The function automatically handles splitting the data into training and testing sets using the specified
partition split. The performance metrics for binary classification problems include AUROC and prAUC. For
multi-class classification, these metrics are not computed, but confusion matrices and post-resampling
statistics are provided.
}
\examples{
\dontrun{
# Example usage:
dataset_ml <- generate_demo_data(n_subjects = 1000, n_features = 200)
selectedPackages <- c("nb", "rpart")
outcome <- "pandora_cluster"
results <- auto_simon_ml(dataset_ml, outcome, selectedPackages, 0.7)
}

}
\keyword{internal}
