% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_tsne}
\alias{calculate_tsne}
\title{Calculate t-SNE}
\usage{
calculate_tsne(dataset, settings, removeGroups = TRUE)
}
\arguments{
\item{dataset}{The dataset to perform the analysis on.}

\item{settings}{A list of settings for the analysis.}

\item{removeGroups}{If TRUE, removes the grouping variables from the dataset.}
}
\value{
A list containing the analyzed data, t-SNE results and column headers.
}
\description{
This function performs a t-SNE analysis on a given dataset.
}
