% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castAllStringsToNA}
\alias{castAllStringsToNA}
\title{Cast All Strings to NA}
\usage{
castAllStringsToNA(dataset, excludeColumns = c())
}
\arguments{
\item{dataset}{Dataframe to be processed}

\item{excludeColumns}{Character vector of column names to exclude}
}
\value{
A dataframe with all strings in non-excluded columns cast to NA
}
\description{
Removes all strings/words from specified columns in a dataset
}
