% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cluster_tsne_mclust}
\alias{cluster_tsne_mclust}
\title{Mclust clustering on t-SNE results}
\usage{
cluster_tsne_mclust(info.norm, tsne.norm, settings)
}
\arguments{
\item{info.norm}{The normalized data on which the t-SNE analysis was carried out.}

\item{tsne.norm}{The t-SNE results.}

\item{settings}{A list of settings for the analysis.}
}
\value{
A list containing the analyzed data with cluster assignment and cluster centroids.
}
\description{
This function applies Mclust clustering on t-SNE results.
}
