% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunaut.R
\name{immunaut}
\alias{immunaut}
\title{Main function to carry out immunaut analysis}
\usage{
immunaut(dataset, settings = list())
}
\arguments{
\item{dataset}{A data frame representing the dataset on which the analysis will be performed.}

\item{settings}{A named list containing the settings for the analysis. If left NULL, defaults will be used.}
}
\value{
The final dataset having an additional column for cluster assignments. If NULL, warning messages would
be printed for any erroneous input parameters.
}
\description{
This function performs different types of clustering on given dataset with user-defined
settings and generates associated plots. It performs error checking, pre-processing,
dimension reduction via t-SNE, clustering, and data visualization in sequence.
}
