% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immunautDemo-data.R
\docType{data}
\name{immunautDemo}
\alias{immunautDemo}
\title{Demo data set from immunaut package.
This data is used in this package examples. It consist of 4x4 feature matrix + additional dummy columns that can be used for testing.}
\format{
An object of class \code{data.frame} with 4 rows and 7 columns.
}
\usage{
data(immunautDemo)
}
\description{
Demo data set from immunaut package.
This data is used in this package examples. It consist of 4x4 feature matrix + additional dummy columns that can be used for testing.
}
\examples{
data(immunautDemo)
print(head(immunautDemo))
results <- immunaut(immunautDemo, settings)

}
\keyword{datasets}
