% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_var_empty}
\alias{is_var_empty}
\title{Check if Request Variable is Empty}
\usage{
is_var_empty(variable)
}
\arguments{
\item{variable}{The input variable to check.}
}
\value{
A boolean value indicating whether the variable is empty.
}
\description{
This function checks whether a given variable is empty. It returns TRUE if the variable is empty; otherwise it returns FALSE.
}
