% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preProcessData}
\alias{preProcessData}
\title{Preprocess a Dataset}
\usage{
preProcessData(data, outcome, excludeClasses, methods = c("center", "scale"))
}
\arguments{
\item{data}{The dataset to be preprocessed.}

\item{outcome}{The outcome variable in the dataset (if any).}

\item{excludeClasses}{A character vector of column names to exclude from preprocessing.}

\item{methods}{A character vector specifying the preprocessing methods to apply.}
}
\value{
A list containing the processed dataset and the preprocessing parameters applied.
}
\description{
This function preprocesses a dataset using a given method such as
centering, scaling, etc., and excludes certain classes if specified.
}
