% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{preProcessResample}
\alias{preProcessResample}
\title{Pre-process and Resample Dataset}
\usage{
preProcessResample(
  datasetData,
  preProcess,
  selectedOutcomeColumns,
  outcome_and_classes
)
}
\arguments{
\item{datasetData}{Dataframe to be pre-processed}

\item{preProcess}{Vector of pre-processing methods to apply}

\item{selectedOutcomeColumns}{Character vector of outcome columns}

\item{outcome_and_classes}{List of outcomes and their classes}
}
\value{
A list containing the pre-processing mapping and the processed dataset
}
\description{
This function applies pre-processing transformations to the dataset, then resamples it.
}
